feat(id): move function to its own `/utils/Identifier.js` file
For ease of maintenance
